refactor: let comments own the reset service
Everything the service does reaches into comments: it clears them and
records the reset. Holding it in the shared services package made that
package depend on a feature package, which is the wrong direction.
