+++ mpvqc/importing/services/importer.py
@@ -10,9 +10,9 @@ from typing import TYPE_CHECKING, Protocol
 import inject
 from PySide6.QtCore import QObject, Signal
 
-from mpvqc.comments.services import CommentsService
+from mpvqc.comments.services import CommentsService, ResetService
 from mpvqc.jobs import Err, Ok, SerialJobRunner
-from mpvqc.services import PlayerService, ResetService, StateService
+from mpvqc.services import PlayerService, StateService
 
 from .concerns import SessionMerge, SessionReplace, SubtitlesLoad, SubtitlesSkip, VideoLoad, VideoSkip
 from .pending import PendingImport
