+++ pyproject.toml
@@ -147,6 +147,7 @@ files = [
     "mpvqc/comments/services/__init__.py",
     "mpvqc/comments/services/commands.py",
     "mpvqc/comments/services/history.py",
+    "mpvqc/comments/services/resetter.py",
     "mpvqc/comments/services/reverse_translator.py",
     "mpvqc/comments/services/roles.py",
     "mpvqc/comments/services/search.py",
@@ -274,7 +275,6 @@ files = [
     "mpvqc/services/player/service.py",
     "mpvqc/services/player/state.py",
     "mpvqc/services/quit.py",
-    "mpvqc/services/resetter.py",
     "mpvqc/services/resource.py",
     "mpvqc/services/settings.py",
     "mpvqc/services/settings_file.py",
